Last changed
Now you can add new Comics as Administrator and see them in the Home
Page.
Also the visualization by genre is dynamic using the content of the
comics genres

+++ js/initialization.js
@@ -18,14 +18,6 @@ if(!existUsrs)
 
 console.log('adding users...');
 
-//roles
-var geners=[{'gnrId':1,'gnrName':'mutants','gnrDesc':'Comics about mutants'},
-			 {'gnrId':2,'gnrName':'superheroes','gnrDesc':'Comics about super heroes'},
-			 {'gnrId':3,'gnrName':'aliens','gnrDesc':'Comics about aliens'},
-			 {'gnrId':4,'gnrName':'zombies','gnrDesc':'Comics about zombies'}];
-localStorage.setItem('geners', JSON.stringify(geners));
-console.log('adding geners...');
-
 //rate
 var rates=[{'rateId':1,'rateName':'5 stars','rateValue':5},
 			{'rateId':2,'rateName':'4 stars','rateValue':4},
@@ -35,32 +27,30 @@ var rates=[{'rateId':1,'rateName':'5 stars','rateValue':5},
 localStorage.setItem('rates', JSON.stringify(rates));
 console.log('adding rates....');
 
-//editions
-var editions=[{'editionsId':1,'editionsName':'first','editionsDesc':'First edition'},
-			{'editionsId':2,'editionsName':'special','editionsDesc':'Special edition'}];
-localStorage.setItem('editions', JSON.stringify(editions));
-console.log('adding editions....');
-
 //comics
+var existComics= JSON.parse(localStorage.getItem('comics'));
+if(!existComics)
+{
 var comics=[
-			{'name':'Spring issue','creator':'Grant Morrison','date':'1997/01/02','edition':'first','genre':'superheroes','mainCharacter':'Batman','numberOfVisits':5,'coverImg':'batman01.jpg','rate':5,'isRecommend':1},
-			{'name':'Number 52','creator':'Grant Morrison','date':'2000/03/03','edition':'first','genre':'superheroes','mainCharacter':'Batman','numberOfVisits':2,'coverImg':'batman02.jpg','rate':4,'isRecommend':0},
-			{'name':'Final act','creator':'Grant Morrison','date':'2006/04/05','edition':'first','genre':'superheroes','mainCharacter':'Batman','numberOfVisits':1,'coverImg':'batman03.jpg','rate':3,'isRecommend':1},
-			{'name':'Chapter 204','creator':'Stan Lee','date':'1980/05/06','edition':'first','genre':'superheroes','mainCharacter':'Superman','numberOfVisits':10,'coverImg':'superman01.jpg','rate':2,'isRecommend':0},
-			{'name':'Chapter 707','creator':'Stan Lee','date':'1990/09/09','edition':'special','genre':'superheroes','mainCharacter':'Superman','numberOfVisits':2,'coverImg':'superman02.jpg','rate':1,'isRecommend':1},
-			{'name':'Chapter 01','creator':'Stan Lee','date':'2005/03/11','edition':'first','genre':'superheroes','mainCharacter':'Superman','numberOfVisits':1,'coverImg':'superman03.jpg','rate':3,'isRecommend':1},
-			{'name':'Dragon ball Z 8','creator':'Akira Toriyama','date':'1983/03/03','edition':'first','genre':'aliens','mainCharacter':'Goku','numberOfVisits':9,'coverImg':'dragonballZ01.jpg','rate':5,'isRecommend':1},
-			{'name':'Dragon ball Z 9','creator':'Akira Toriyama','date':'1984/04/05','edition':'first','genre':'aliens','mainCharacter':'Goku','numberOfVisits':4,'coverImg':'dragonballZ02.jpg','rate':4,'isRecommend':1},
-			{'name':'Dragon ball Z 12','creator':'Akira Toriyama','date':'1985/09/12','edition':'first','genre':'aliens','mainCharacter':'Goku','numberOfVisits':1,'coverImg':'dragonballZ03.jpg','rate':1,'isRecommend':1},
-			{'name':'Chapter 17','creator':'Aaron Palo','date':'2004/01/01','edition':'first','genre':'mutants','mainCharacter':'Wolverine','numberOfVisits':11,'coverImg':'wolverine01.jpg','rate':3,'isRecommend':1},
-			{'name':'Chapter 004','creator':'Aaron Palo','date':'2005/12/10/','edition':'first','genre':'mutants','mainCharacter':'Wolverine','numberOfVisits':8,'coverImg':'wolverine02.jpg','rate':2,'isRecommend':1},
-			{'name':'Chapter5','creator':'Aaron Palo','date':'2006/04/02','edition':'first','genre':'mutants','mainCharacter':'Wolverine','numberOfVisits':3,'coverImg':'wolverine03.jpg','rate':1,'isRecommend':0},
-			{'name':'Birth of the power','creator':'Stan Lee','date':'1988/03/08','edition':'first','genre':'superheroes','mainCharacter':'Iron Man','numberOfVisits':7,'coverImg':'ironman01.jpg','rate':5,'isRecommend':1},
-			{'name':'Hammer','creator':'Stan Lee','date':'1989/05/22/','edition':'special','genre':'superheroes','mainCharacter':'Iron Man','numberOfVisits':4,'coverImg':'ironman02.jpg','rate':4,'isRecommend':0},
-			{'name':'Legacy','creator':'Stan Lee','date':'1990/12/24','edition':'first','genre':'superheroes','mainCharacter':'Iron Man','numberOfVisits':1,'coverImg':'ironman03.jpg','rate':2,'isRecommend':0},
-			{'name':'Chapter 1','creator':'Robert Kirkman','date':'2005/02/13','edition':'first','genre':'zombies','mainCharacter':'Rick Grames','numberOfVisits':14,'coverImg':'walkingdead01.jpg','rate':5,'isRecommend':1},
-			{'name':'Chapter 2','creator':'Robert Kirkman','date':'2006/02/14','edition':'first','genre':'zombies','mainCharacter':'Rick Grames','numberOfVisits':6,'coverImg':'walkingdead02.jpg','rate':5,'isRecommend':1},
-			{'name':'Chapter 100','creator':'Robert Kirkman','date':'2007/02/16','edition':'first','genre':'zombies','mainCharacter':'Rick Grames','numberOfVisits':2,'coverImg':'walkingdead03.jpg','rate':5,'isRecommend':1}
+			{'name':'Spring issue','creator':'Grant Morrison','date':'1997/01/02','edition':'first','genre':'superheroes','mainCharacter':'Batman','numberOfVisits':5,'coverImg':'batman01.jpg','rate':5,'isRecommend':1,'searches':245},
+			{'name':'Number 52','creator':'Grant Morrison','date':'2000/03/03','edition':'first','genre':'superheroes','mainCharacter':'Batman','numberOfVisits':2,'coverImg':'batman02.jpg','rate':4,'isRecommend':0,'searches':315},
+			{'name':'Final act','creator':'Grant Morrison','date':'2006/04/05','edition':'first','genre':'superheroes','mainCharacter':'Batman','numberOfVisits':1,'coverImg':'batman03.jpg','rate':3,'isRecommend':1,'searches':824},
+			{'name':'Chapter 204','creator':'Stan Lee','date':'1980/05/06','edition':'first','genre':'superheroes','mainCharacter':'Superman','numberOfVisits':10,'coverImg':'superman01.jpg','rate':2,'isRecommend':0,'searches':445},
+			{'name':'Chapter 707','creator':'Stan Lee','date':'1990/09/09','edition':'special','genre':'superheroes','mainCharacter':'Superman','numberOfVisits':2,'coverImg':'superman02.jpg','rate':1,'isRecommend':1,'searches':20},
+			{'name':'Chapter 01','creator':'Stan Lee','date':'2005/03/11','edition':'first','genre':'superheroes','mainCharacter':'Superman','numberOfVisits':1,'coverImg':'superman03.jpg','rate':3,'isRecommend':1,'searches':397},
+			{'name':'Dragon ball Z 8','creator':'Akira Toriyama','date':'1983/03/03','edition':'first','genre':'aliens','mainCharacter':'Goku','numberOfVisits':9,'coverImg':'dragonballZ01.jpg','rate':5,'isRecommend':1,'searches':417},
+			{'name':'Dragon ball Z 9','creator':'Akira Toriyama','date':'1984/04/05','edition':'first','genre':'aliens','mainCharacter':'Goku','numberOfVisits':4,'coverImg':'dragonballZ02.jpg','rate':4,'isRecommend':1,'searches':987},
+			{'name':'Dragon ball Z 12','creator':'Akira Toriyama','date':'1985/09/12','edition':'first','genre':'aliens','mainCharacter':'Goku','numberOfVisits':1,'coverImg':'dragonballZ03.jpg','rate':1,'isRecommend':1,'searches':845},
+			{'name':'Chapter 17','creator':'Aaron Palo','date':'2004/01/01','edition':'first','genre':'mutants','mainCharacter':'Wolverine','numberOfVisits':11,'coverImg':'wolverine01.jpg','rate':3,'isRecommend':1,'searches':674},
+			{'name':'Chapter 004','creator':'Aaron Palo','date':'2005/12/10/','edition':'first','genre':'mutants','mainCharacter':'Wolverine','numberOfVisits':8,'coverImg':'wolverine02.jpg','rate':2,'isRecommend':1,'searches':12},
+			{'name':'Chapter5','creator':'Aaron Palo','date':'2006/04/02','edition':'first','genre':'mutants','mainCharacter':'Wolverine','numberOfVisits':3,'coverImg':'wolverine03.jpg','rate':1,'isRecommend':0,'searches':87},
+			{'name':'Birth of the power','creator':'Stan Lee','date':'1988/03/08','edition':'first','genre':'superheroes','mainCharacter':'Iron Man','numberOfVisits':7,'coverImg':'ironman01.jpg','rate':5,'isRecommend':1,'searches':99},
+			{'name':'Hammer','creator':'Stan Lee','date':'1989/05/22/','edition':'special','genre':'superheroes','mainCharacter':'Iron Man','numberOfVisits':4,'coverImg':'ironman02.jpg','rate':4,'isRecommend':0,'searches':584},
+			{'name':'Legacy','creator':'Stan Lee','date':'1990/12/24','edition':'first','genre':'superheroes','mainCharacter':'Iron Man','numberOfVisits':1,'coverImg':'ironman03.jpg','rate':2,'isRecommend':0,'searches':125},
+			{'name':'Chapter 1','creator':'Robert Kirkman','date':'2005/02/13','edition':'first','genre':'zombies','mainCharacter':'Rick Grames','numberOfVisits':14,'coverImg':'walkingdead01.jpg','rate':5,'isRecommend':1,'searches':133},
+			{'name':'Chapter 2','creator':'Robert Kirkman','date':'2006/02/14','edition':'first','genre':'zombies','mainCharacter':'Rick Grames','numberOfVisits':6,'coverImg':'walkingdead02.jpg','rate':5,'isRecommend':1,'searches':445},
+			{'name':'Chapter 100','creator':'Robert Kirkman','date':'2007/02/16','edition':'first','genre':'zombies','mainCharacter':'Rick Grames','numberOfVisits':2,'coverImg':'walkingdead03.jpg','rate':5,'isRecommend':1,'searches':704}
 			];
 			localStorage.setItem('comics', JSON.stringify(comics));
+}
 console.log('adding comics....');
